Initial code for dropping blocks in hurry mode.
(Commented out because it isn't ready.)

The idea is that they'll not be solid when landing, but detect if they overlap the player (to kill him). When they get to their destination, stop detecting overlap and turn on solid.

Also need to code the loop to wrap around the perimeter.