Added the falling bricks when time is running out.
Had to make a remove() method for bombs, where they can be destroyed without the explosion so the bomb cnt can be attributed back to the player.

Refactored a canDie component so player can more easily be triggered for its death. Will especially be useful when we add more sophisticated death sequence.

Some of this can probably be factored into a Block component but for now its in gameScene.