Add support for P-Bomb powerup.
Detects double-space by tracking the last time and position when space was pressed. If within 0.3 seconds and position hasn't changed, it is double-space.

Refactor spawnBomb to use spawnBombAtPos, which does the heavy lifting, and returns a boolean whether it succeeded. For instance, it fails if the player can't spawn a bomb (no more bombs) or if something is blocking the bomb at that position.

Then, spawnPBomb uses the same code but with a while loop. Works great!