Fix blocks to correctly squash enemies/players.
The bug was that they wouldn't squash a player or enemy unless it was exactly in the right snapped grid position, which almost never happens. So the logic was improved to check that the block was overlapping with the area of items, instead of checking a snapped coordinate point.

This logic can probably be used at some other places to improve collision checking.