Set up a "Main Menu" screen separate from start screen.
So the "Start" screen is where the game starts, and is used to force user interaction so the main menu music can kick in.

The "Main Menu" screen is then the main menu you expect at the beginning.

NOTE: There is some non-DRY code here. The start and main menu scenes both have the game title/logo effect.