Initial work on a player selection scene.
It allows number of players, and that is passed on to the game scene
but that's it.

Still need to allow actual character selection and which control is for
which character.

Also allows you to enter an online room name although it doesn't
actually connect online yet. Shows an error message indicating it isn't
available yet.

Also: Increased speed of the hurry music.